starting pipes

+++ activitats/act05/player.c
@@ -3,24 +3,25 @@
 void signalHandler(int sig);
 
 int main(int argc, char *argv[]) {
-    
-    for(int i = 0; i < 10; i++) {
-        int key = generatePassword();
-        printf("[Child: %d] Password: %04d\n", getpid(), key);
-        //write(fd[1], key, PASS_LENGTH);
+
+    if(argc != 2) {
+        exit(1);
+    }
+
+    char *code = argv[1];
+
+    while(1) {
+        char key[PASS_LENGTH];
+        generatePassword(key);
+
+        if (strncmp(key, code, PASS_LENGTH) == 0) {
+            kill(getppid(), SIGUSR2);
+        }
+
         //sleep 500ms
-        usleep(500 * 1000);
+        usleep(100 * 1000);
 
     }
-
-    kill(getppid(), SIGUSR2);
-
+    
     return 0;
-}
-
-void signalHandler(int sig) {
-    if (sig == SIGUSR2) {
-        printf("[Child %d]Time's up!\n", getpid());
-        exit(0);
-    }
 }
